REFACTOR : Refactoring navbar logic with composable #39
