+++ src/composables/index.ts
@@ -2,6 +2,7 @@ import { useToolTipGraph } from './graph/toolTipGraph';
 import { useBuildGraph } from './graph/buildGraph';
 import { useSetGraph } from './graph/setGraph';
 import { useScreen } from "./screen/screen";
+import { useNavbar } from "./screen/navbar"
 import { useAuth } from "./auth/auth";
 
-export { useBuildGraph, useSetGraph, useScreen, useAuth, useToolTipGraph };
+export { useBuildGraph, useSetGraph, useScreen, useAuth, useToolTipGraph, useNavbar };
